requirements.txt
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.1
click==8.1.3
fastapi==0.82.0
FastAPI-SQLAlchemy==0.2.1
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2==2.9.5
pydantic==1.10.2
pypika-tortoise==0.1.6
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlmodel==0.0.8
starlette==0.19.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
